refactor(net)!: `ariel_os_embassy::network` -> `ariel_os_embassy::net`
